feat: file extension of ES modules will be .mjs (#734)
If the compiled catalog is an ES module then the file extension will be `.mjs`. This is good for Node.js which can import ES modules natively, but the modules must be in a .mjs file. For users of module bunlders, like Webpack, have to test for .mjs files, like `test: /\.m?js$/`