Explain config in the Setup tutorial
A new user has no idea how the configuration file should be, so writing "After fixing configuration" in the tutorial is not helping anything.
And actually just putting the locales in the config will make lingui scan the /node_modules/ directory, which will throw a lot of errors. 
So a minimal working .linguirc/config example would be great.