feat(macro): Add support for passing custom i18n instance
This allows developers that work in concurrent rendering environments
to use a Lingui instance that is locale-specific, without risking global
Lingui to be switched to another locale.

In React, `<I18nProvider>` may be used to provide a locale-specific
Lingui instance, and it can be retrieved by using `useLingui`. After
that, the instance can be passed as argument to the `t` macro and
it will be used instead of the global instance, ensuring deterministic
behaviour.