Remove the obsolete `defineMessages` function from the typings.
When working in a TypeScript environment, the typings suggest that a function called `defineMessages` (note the plural) should be available. However, this function is not actually available in the code, and its use will cause errors at runtime. It appears that this function was removed as obsolete a while back, so it should also be removed from the typings.

The only other place this function was referenced was in the switch statement in `getMacroType()`, which is also cleaned up in this change.