fix(cli): use caret range for `micromatch` dependency
4.0.2 contains a vulnerability that most likely doesn't affect @lingui/cli, but there should be no
reason to depend on an exact version which prevents consumers from upgrading the package without a
version conflict.
I've kept the version at 4.0.2 to prevent a breaking change (in the sense that bumping it to ^4.0.8
might also cause a version conflict for consumers). Let me know if you'd rather bump it to ^4.0.8.
Ref: https://github.com/advisories/GHSA-952p-6rrq-rcjv